Added IPropertySet support for creating EGL surfaces with custom properties.

+++ src/common/winrt/winrtutils.h
@@ -15,7 +15,7 @@ namespace winrt
 {
 
 std::string getTempPath();
-float convertDipsToPixels(float dips);
+long convertDipsToPixels(float dips);
 
 };
